+++ resources/release-notes.md
@@ -1,4 +1,5 @@
 # Software
 
 - Add separate figure in `mag.graphics.view.Frequency` to show field and spectrogram
+- Add `mag.time.Constant.Eps` for consistent for consistent definition of `eps` in seconds
 - Detect mode changes from timestamp cadence, when no event data is available